fix Testbed build

+++ Box2D/Testbed/Framework/DebugDraw.cpp
@@ -32,12 +32,12 @@ Camera g_camera;
 //
 b2Vec2 Camera::ConvertScreenToWorld(const b2Vec2& ps)
 {
-    float32 w = float32(m_width);
-    float32 h = float32(m_height);
-	float32 u = ps.x / w;
-	float32 v = (h - ps.y) / h;
+    double w = double(m_width);
+    double h = double(m_height);
+	double u = ps.x / w;
+	double v = (h - ps.y) / h;
 
-	float32 ratio = w / h;
+	double ratio = w / h;
 	b2Vec2 extents(ratio * 25.0f, 25.0f);
 	extents *= m_zoom;
 
@@ -53,17 +53,17 @@ b2Vec2 Camera::ConvertScreenToWorld(const b2Vec2& ps)
 //
 b2Vec2 Camera::ConvertWorldToScreen(const b2Vec2& pw)
 {
-	float32 w = float32(m_width);
-	float32 h = float32(m_height);
-	float32 ratio = w / h;
+	double w = double(m_width);
+	double h = double(m_height);
+	double ratio = w / h;
 	b2Vec2 extents(ratio * 25.0f, 25.0f);
 	extents *= m_zoom;
 
 	b2Vec2 lower = m_center - extents;
 	b2Vec2 upper = m_center + extents;
 
-	float32 u = (pw.x - lower.x) / (upper.x - lower.x);
-	float32 v = (pw.y - lower.y) / (upper.y - lower.y);
+	double u = (pw.x - lower.x) / (upper.x - lower.x);
+	double v = (pw.y - lower.y) / (upper.y - lower.y);
 
 	b2Vec2 ps;
 	ps.x = u * w;
@@ -73,11 +73,11 @@ b2Vec2 Camera::ConvertWorldToScreen(const b2Vec2& pw)
 
 // Convert from world coordinates to normalized device coordinates.
 // http://www.songho.ca/opengl/gl_projectionmatrix.html
-void Camera::BuildProjectionMatrix(float32* m, float32 zBias)
+void Camera::BuildProjectionMatrix(float* m, float zBias)
 {
-	float32 w = float32(m_width);
-	float32 h = float32(m_height);
-	float32 ratio = w / h;
+	double w = double(m_width);
+	double h = double(m_height);
+	double ratio = w / h;
 	b2Vec2 extents(ratio * 25.0f, 25.0f);
 	extents *= m_zoom;
 
@@ -252,7 +252,7 @@ struct GLRenderPoints
 		}
 	}
     
-	void Vertex(const b2Vec2& v, const b2Color& c, float32 size)
+	void Vertex(const b2Vec2& v, const b2Color& c, double size)
 	{
 		if (m_count == e_maxVertices)
 			Flush();
@@ -271,7 +271,7 @@ struct GLRenderPoints
         
 		glUseProgram(m_programId);
         
-		float32 proj[16] = { 0.0f };
+        float proj[16] = { 0.0f };
 		g_camera.BuildProjectionMatrix(proj, 0.0f);
         
 		glUniformMatrix4fv(m_projectionUniform, 1, GL_FALSE, proj);
@@ -305,7 +305,7 @@ struct GLRenderPoints
 	enum { e_maxVertices = 512 };
 	struct Vertex m_vertices[e_maxVertices];
 
-	int32 m_count;
+	int32_t m_count;
     
 	GLuint m_vboId;
 	GLuint m_programId;
@@ -397,7 +397,7 @@ struct GLRenderLines
         
 		glUseProgram(m_programId);
         
-		float32 proj[16] = { 0.0f };
+        float proj[16] = { 0.0f };
 		g_camera.BuildProjectionMatrix(proj, 0.1f);
         
 		glUniformMatrix4fv(m_projectionUniform, 1, GL_FALSE, proj);
@@ -426,7 +426,7 @@ struct GLRenderLines
 	enum { e_maxVertices = 2 * 512 };
 	struct Vertex m_vertices[e_maxVertices];
     
-	int32 m_count;
+	int32_t m_count;
     
 	GLuint m_vboId;
 	GLuint m_programId;
@@ -517,7 +517,7 @@ struct GLRenderTriangles
         
 		glUseProgram(m_programId);
         
-		float32 proj[16] = { 0.0f };
+        float proj[16] = { 0.0f };
 		g_camera.BuildProjectionMatrix(proj, 0.2f);
         
 		glUniformMatrix4fv(m_projectionUniform, 1, GL_FALSE, proj);
@@ -549,7 +549,7 @@ struct GLRenderTriangles
 	enum { e_maxVertices = 3 * 512 };
 	struct Vertex m_vertices[e_maxVertices];
 
-	int32 m_count;
+	int32_t m_count;
 
 	GLuint m_vboId;
 	GLuint m_programId;
@@ -602,10 +602,10 @@ void DebugDraw::Destroy()
 }
 
 //
-void DebugDraw::DrawPolygon(const b2Vec2* vertices, int32 vertexCount, const b2Color& color)
+void DebugDraw::DrawPolygon(const b2Vec2* vertices, int32_t vertexCount, const b2Color& color)
 {
     b2Vec2 p1 = vertices[vertexCount - 1];
-	for (int32 i = 0; i < vertexCount; ++i)
+	for (int32_t i = 0; i < vertexCount; ++i)
 	{
         b2Vec2 p2 = vertices[i];
 		m_lines->Vertex(p1, color);
@@ -615,11 +615,11 @@ void DebugDraw::DrawPolygon(const b2Vec2* vertices, int32 vertexCount, const b2C
 }
 
 //
-void DebugDraw::DrawSolidPolygon(const b2Vec2* vertices, int32 vertexCount, const b2Color& color)
+void DebugDraw::DrawSolidPolygon(const b2Vec2* vertices, int32_t vertexCount, const b2Color& color)
 {
 	b2Color fillColor(0.5f * color.r, 0.5f * color.g, 0.5f * color.b, 0.5f);
 
-    for (int32 i = 1; i < vertexCount - 1; ++i)
+    for (int32_t i = 1; i < vertexCount - 1; ++i)
     {
         m_triangles->Vertex(vertices[0], fillColor);
         m_triangles->Vertex(vertices[i], fillColor);
@@ -627,7 +627,7 @@ void DebugDraw::DrawSolidPolygon(const b2Vec2* vertices, int32 vertexCount, cons
     }
 
     b2Vec2 p1 = vertices[vertexCount - 1];
-	for (int32 i = 0; i < vertexCount; ++i)
+	for (int32_t i = 0; i < vertexCount; ++i)
 	{
         b2Vec2 p2 = vertices[i];
 		m_lines->Vertex(p1, color);
@@ -637,15 +637,15 @@ void DebugDraw::DrawSolidPolygon(const b2Vec2* vertices, int32 vertexCount, cons
 }
 
 //
-void DebugDraw::DrawCircle(const b2Vec2& center, float32 radius, const b2Color& color)
+void DebugDraw::DrawCircle(const b2Vec2& center, double radius, const b2Color& color)
 {
-	const float32 k_segments = 16.0f;
-	const float32 k_increment = 2.0f * b2_pi / k_segments;
-    float32 sinInc = sinf(k_increment);
-    float32 cosInc = cosf(k_increment);
+	const double k_segments = 16.0f;
+	const double k_increment = 2.0f * b2_pi / k_segments;
+    double sinInc = sinf(k_increment);
+    double cosInc = cosf(k_increment);
     b2Vec2 r1(1.0f, 0.0f);
     b2Vec2 v1 = center + radius * r1;
-	for (int32 i = 0; i < k_segments; ++i)
+	for (int32_t i = 0; i < k_segments; ++i)
 	{
         // Perform rotation to avoid additional trigonometry.
         b2Vec2 r2;
@@ -660,17 +660,17 @@ void DebugDraw::DrawCircle(const b2Vec2& center, float32 radius, const b2Color& 
 }
 
 //
-void DebugDraw::DrawSolidCircle(const b2Vec2& center, float32 radius, const b2Vec2& axis, const b2Color& color)
+void DebugDraw::DrawSolidCircle(const b2Vec2& center, double radius, const b2Vec2& axis, const b2Color& color)
 {
-	const float32 k_segments = 16.0f;
-	const float32 k_increment = 2.0f * b2_pi / k_segments;
-    float32 sinInc = sinf(k_increment);
-    float32 cosInc = cosf(k_increment);
+	const double k_segments = 16.0f;
+	const double k_increment = 2.0f * b2_pi / k_segments;
+    double sinInc = sinf(k_increment);
+    double cosInc = cosf(k_increment);
     b2Vec2 v0 = center;
     b2Vec2 r1(cosInc, sinInc);
     b2Vec2 v1 = center + radius * r1;
 	b2Color fillColor(0.5f * color.r, 0.5f * color.g, 0.5f * color.b, 0.5f);
-	for (int32 i = 0; i < k_segments; ++i)
+	for (int32_t i = 0; i < k_segments; ++i)
 	{
         // Perform rotation to avoid additional trigonometry.
         b2Vec2 r2;
@@ -686,7 +686,7 @@ void DebugDraw::DrawSolidCircle(const b2Vec2& center, float32 radius, const b2Ve
 
     r1.Set(1.0f, 0.0f);
     v1 = center + radius * r1;
-	for (int32 i = 0; i < k_segments; ++i)
+	for (int32_t i = 0; i < k_segments; ++i)
 	{
         b2Vec2 r2;
         r2.x = cosInc * r1.x - sinInc * r1.y;
@@ -714,7 +714,7 @@ void DebugDraw::DrawSegment(const b2Vec2& p1, const b2Vec2& p2, const b2Color& c
 //
 void DebugDraw::DrawTransform(const b2Transform& xf)
 {
-	const float32 k_axisScale = 0.4f;
+	const double k_axisScale = 0.4f;
     b2Color red(1.0f, 0.0f, 0.0f);
     b2Color green(0.0f, 1.0f, 0.0f);
 	b2Vec2 p1 = xf.p, p2;
@@ -728,7 +728,7 @@ void DebugDraw::DrawTransform(const b2Transform& xf)
 	m_lines->Vertex(p2, green);
 }
 
-void DebugDraw::DrawPoint(const b2Vec2& p, float32 size, const b2Color& color)
+void DebugDraw::DrawPoint(const b2Vec2& p, double size, const b2Color& color)
 {
     m_points->Vertex(p, color, size);
 }
