+++ Box2D/Testbed/Framework/Main.cpp
@@ -43,9 +43,9 @@ namespace
 	GLFWwindow* mainWindow = NULL;
 	UIState ui;
 
-	int32 testIndex = 0;
-	int32 testSelection = 0;
-	int32 testCount = 0;
+	int32_t testIndex = 0;
+	int32_t testSelection = 0;
+	int32_t testCount = 0;
 	TestEntry* entry;
 	Test* test;
 	Settings settings;
@@ -203,13 +203,13 @@ static void sKeyCallback(GLFWwindow* window, int key, int scancode, int action, 
 }
 
 //
-static void sMouseButton(GLFWwindow* window, int32 button, int32 action, int32 mods)
+static void sMouseButton(GLFWwindow* window, int32_t button, int32_t action, int32_t mods)
 {
 	ImGui_ImplGlfw_MouseButtonCallback(window, button, action, mods);
 
 	double xd, yd;
 	glfwGetCursorPos(mainWindow, &xd, &yd);
-	b2Vec2 ps((float32)xd, (float32)yd);
+	b2Vec2 ps((double)xd, (double)yd);
 
 	// Use the mouse to move things around.
 	if (button == GLFW_MOUSE_BUTTON_1)
