+++ Box2D/Testbed/Framework/Test.h
@@ -36,18 +36,18 @@ typedef Test* TestCreateFcn();
 #define DRAW_STRING_NEW_LINE 16
 
 /// Random number in range [-1,1]
-inline float32 RandomFloat()
+inline double RandomFloat()
 {
-	float32 r = (float32)(rand() & (RAND_LIMIT));
+	double r = (double)(rand() & (RAND_LIMIT));
 	r /= RAND_LIMIT;
 	r = 2.0f * r - 1.0f;
 	return r;
 }
 
 /// Random floating point number in range [lo, hi]
-inline float32 RandomFloat(float32 lo, float32 hi)
+inline double RandomFloat(double lo, double hi)
 {
-	float32 r = (float32)(rand() & (RAND_LIMIT));
+	double r = (double)(rand() & (RAND_LIMIT));
 	r /= RAND_LIMIT;
 	r = (hi - lo) * r + lo;
 	return r;
@@ -79,9 +79,9 @@ struct Settings
 		singleStep = false;
 	}
 
-	float32 hz;
-	int32 velocityIterations;
-	int32 positionIterations;
+    float hz;
+	int32_t velocityIterations;
+	int32_t positionIterations;
 	bool drawShapes;
 	bool drawJoints;
 	bool drawAABBs;
@@ -119,7 +119,7 @@ public:
 	Test* test;
 };
 
-const int32 k_maxContactPoints = 2048;
+const int32_t k_maxContactPoints = 2048;
 
 struct ContactPoint
 {
@@ -128,9 +128,9 @@ struct ContactPoint
 	b2Vec2 normal;
 	b2Vec2 position;
 	b2PointState state;
-	float32 normalImpulse;
-	float32 tangentImpulse;
-	float32 separation;
+	double normalImpulse;
+	double tangentImpulse;
+	double separation;
 };
 
 class Test : public b2ContactListener
@@ -177,16 +177,16 @@ protected:
 	b2Body* m_groundBody;
 	b2AABB m_worldAABB;
 	ContactPoint m_points[k_maxContactPoints];
-	int32 m_pointCount;
+	int32_t m_pointCount;
 	DestructionListener m_destructionListener;
-	int32 m_textLine;
+	int32_t m_textLine;
 	b2World* m_world;
 	b2Body* m_bomb;
 	b2MouseJoint* m_mouseJoint;
 	b2Vec2 m_bombSpawnPoint;
 	bool m_bombSpawning;
 	b2Vec2 m_mouseWorld;
-	int32 m_stepCount;
+	int32_t m_stepCount;
 
 	b2Profile m_maxProfile;
 	b2Profile m_totalProfile;
